feat: Add astro-decap-cms-oauth package
This commit adds the astro-decap-cms-oauth package to the project. The package is imported in the astro.config.mjs file and added as a plugin in the defineConfig function. Additionally, the package is added as a dependency in both the package-lock.json and package.json files.

The version of astro-decap-cms-oauth added is 0.2.4, and it has a peer dependency on astro (version ^2.0.0 || ^3.0.0). The package also has a dependency on tiny-json-http (version ^7.5.1), which is included in the node_modules directory.

These changes enable authentication with Decap CMS using OAuth in the Astro project.